backend/requirements.txt
fastapi
python-multipart
uvicorn
pymupdf
python-pptx
